+++ build_db.py
@@ -16,9 +16,24 @@ if __name__ == '__main__':
     os.makedirs('./data', exist_ok=True)
 
     from flask import Flask
-
     app = Flask(__name__, instance_relative_config=True)
-    app.config.from_pyfile('config.py', silent=True)
+
+    env = os.getenv("ENV", "production")
+    if env == "development":
+        # Load the instance development config
+        app.config.from_object("instance.development.Config")
+    elif env == "staging":
+        # Load the instance staging config
+        app.config.from_object("instance.staging.Config")
+    elif env == "production":
+        # Load the instance production config
+        app.config.from_object("instance.production.Config")
+    elif env == "production-dds":
+        # Load the instance production config for the DDS servers
+        app.config.from_object("instance.production-dds.Config")
+    else:
+        raise RuntimeError("Unknown environment '{}'".format(env))
+
     app.app_context().push()
 
     from queuemanager.database import init_app, init_db