+++ instance/gunicorn.conf.py
@@ -13,10 +13,7 @@ if not os.path.exists(_RUN):
 loglevel = 'info'
 
 # Configure the socket address and the number of workers
-workers = 1
 worker_class = 'eventlet'
 
 timeout = 3 * 60  # 3 minutes
 keepalive = 24 * 60 * 60  # 1 day
-
-capture_output = True