+++ queuemanager/socketio/tests/namespaces/test_printer_namespace.py
@@ -15,13 +15,13 @@ from datetime import timedelta
 from queuemanager.socketio import printer_namespace
 
 
-def test_printer_connected(socketio_printer):
-    received_events = socketio_printer.get_received("/client")
+def test_printer_connected(socketio_printer, socketio_client):
+    received_events = socketio_client.get_received("/client")
 
     assert len(received_events) == 0
 
 
-def test_printer_disconnected(socketio_printer, db_manager):
+def test_printer_disconnected(socketio_printer, socketio_client, db_manager):
     printer = db_manager.get_printers(id=1)
 
     db_manager.update_printer(printer, idState=db_manager.printer_state_ids["Ready"], sid=socketio_printer.sid)
@@ -30,7 +30,7 @@ def test_printer_disconnected(socketio_printer, db_manager):
 
     socketio_printer.disconnect("/printer")
 
-    received_events = socketio_printer.get_received("/client")
+    received_events = socketio_client.get_received("/client")
 
     assert len(received_events) == 2
     assert received_events[0]['name'] == 'printer_data_updated'
@@ -94,7 +94,7 @@ def test_emit_job_recovered(socketio_printer, db_manager):
             assert False
 
 
-def test_on_initial_data(socketio_printer, printer_session_key, db_manager):
+def test_on_initial_data(socketio_printer, socketio_client, printer_session_key, db_manager):
     user = db_manager.get_users(id=1)
     file = db_manager.insert_file(user, "test", "/home/Marc/test")
     job = db_manager.insert_job("test", file, user)
@@ -127,7 +127,7 @@ def test_on_initial_data(socketio_printer, printer_session_key, db_manager):
 
     socketio_printer.emit("initial_data", initial_data, namespace="/printer")
 
-    received_events = socketio_printer.get_received("/client")
+    received_events = socketio_client.get_received("/client")
 
     assert len(received_events) == 4
     assert received_events[0]['name'] == 'printer_data_updated'
@@ -176,7 +176,7 @@ def test_on_initial_data(socketio_printer, printer_session_key, db_manager):
             assert False
 
 
-def test_on_state_updated(socketio_printer, printer_session_key, db_manager):
+def test_on_state_updated(socketio_printer, socketio_client, printer_session_key, db_manager):
     user = db_manager.get_users(id=1)
     file = db_manager.insert_file(user, "test", "/home/Marc/test")
     job = db_manager.insert_job("test", file, user)
@@ -193,7 +193,7 @@ def test_on_state_updated(socketio_printer, printer_session_key, db_manager):
 
     socketio_printer.emit("state_updated", state_data, namespace="/printer")
 
-    received_events = socketio_printer.get_received("/client")
+    received_events = socketio_client.get_received("/client")
 
     assert len(received_events) == 2
     assert received_events[0]['name'] == 'printer_data_updated'
@@ -215,7 +215,7 @@ def test_on_state_updated(socketio_printer, printer_session_key, db_manager):
     assert printer.current_job.id == job.id
 
 
-def test_on_extruders_updated(socketio_printer, printer_session_key, db_manager):
+def test_on_extruders_updated(socketio_printer, socketio_client, printer_session_key, db_manager):
     initial_data = {
         "session_key": printer_session_key,
         "extruders_info": [
@@ -237,7 +237,7 @@ def test_on_extruders_updated(socketio_printer, printer_session_key, db_manager)
 
     socketio_printer.emit("extruders_updated", initial_data, namespace="/printer")
 
-    received_events = socketio_printer.get_received("/client")
+    received_events = socketio_client.get_received("/client")
 
     assert len(received_events) == 2
     assert received_events[0]['name'] == 'printer_data_updated'
@@ -266,7 +266,7 @@ def test_on_extruders_updated(socketio_printer, printer_session_key, db_manager)
             assert False
 
 
-def test_on_print_started(socketio_printer, printer_session_key, db_manager):
+def test_on_print_started(socketio_printer, socketio_client, printer_session_key, db_manager):
     user = db_manager.get_users(id=1)
     file = db_manager.insert_file(user, "test", "/home/Marc/test")
     job = db_manager.insert_job("test", file, user)
@@ -282,7 +282,7 @@ def test_on_print_started(socketio_printer, printer_session_key, db_manager):
 
     socketio_printer.emit("print_started", job_data, namespace="/printer")
 
-    received_events = socketio_printer.get_received("/client")
+    received_events = socketio_client.get_received("/client")
 
     assert len(received_events) == 1
     assert received_events[0]['name'] == 'jobs_updated'
@@ -293,7 +293,7 @@ def test_on_print_started(socketio_printer, printer_session_key, db_manager):
     assert job.state.stateString == "Printing"
 
 
-def test_on_print_finished(socketio_printer, printer_session_key, db_manager):
+def test_on_print_finished(socketio_printer, socketio_client, printer_session_key, db_manager):
     user = db_manager.get_users(id=1)
     file = db_manager.insert_file(user, "test", "/home/Marc/test")
     job = db_manager.insert_job("test", file, user)
@@ -311,7 +311,7 @@ def test_on_print_finished(socketio_printer, printer_session_key, db_manager):
 
     socketio_printer.emit("print_finished", job_data, namespace="/printer")
 
-    received_events = socketio_printer.get_received("/client")
+    received_events = socketio_client.get_received("/client")
 
     assert len(received_events) == 2
     assert received_events[0]['name'] == 'job_progress_updated'
@@ -331,7 +331,7 @@ def test_on_print_finished(socketio_printer, printer_session_key, db_manager):
     assert job.interrupted is True
 
 
-def test_on_print_feedback(socketio_printer, printer_session_key, db_manager):
+def test_on_print_feedback(socketio_printer, socketio_client, printer_session_key, db_manager):
     user = db_manager.get_users(id=1)
     file = db_manager.insert_file(user, "test", "/home/Marc/test")
     job = db_manager.insert_job("test", file, user)
@@ -354,7 +354,7 @@ def test_on_print_feedback(socketio_printer, printer_session_key, db_manager):
 
     socketio_printer.emit("print_feedback", feedback_data, namespace="/printer")
 
-    received_events = socketio_printer.get_received("/client")
+    received_events = socketio_client.get_received("/client")
 
     assert len(received_events) == 2
     assert received_events[0]['name'] == 'printer_data_updated'
@@ -369,7 +369,7 @@ def test_on_print_feedback(socketio_printer, printer_session_key, db_manager):
     assert job.state.stateString == "Done"
 
 
-def test_on_printer_temperatures_updated(socketio_printer, printer_session_key, db_manager):
+def test_on_printer_temperatures_updated(socketio_printer, socketio_client, printer_session_key, db_manager):
     temp_data = {
         "bed_temp": 55.1,
         "extruders_temp": [
@@ -388,14 +388,14 @@ def test_on_printer_temperatures_updated(socketio_printer, printer_session_key, 
 
     socketio_printer.emit("printer_temperatures_updated", data_to_send, namespace="/printer")
 
-    received_events = socketio_printer.get_received("/client")
+    received_events = socketio_client.get_received("/client")
 
     assert len(received_events) == 1
     assert received_events[0]['name'] == 'printer_temperatures_updated'
     assert received_events[0]['args'][0] == temp_data
 
 
-def test_on_job_progress_updated(socketio_printer, printer_session_key, db_manager):
+def test_on_job_progress_updated(socketio_printer, socketio_client, printer_session_key, db_manager):
     user = db_manager.get_users(id=1)
     file = db_manager.insert_file(user, "test", "/home/Marc/test")
     job = db_manager.insert_job("test", file, user)
@@ -416,7 +416,7 @@ def test_on_job_progress_updated(socketio_printer, printer_session_key, db_manag
 
     socketio_printer.emit("job_progress_updated", data_to_send, namespace="/printer")
 
-    received_events = socketio_printer.get_received("/client")
+    received_events = socketio_client.get_received("/client")
 
     expected_progress_data = progress_data.copy()
     expected_progress_data["name"] = "test"